+++ api/app/Http/Controllers/OrgController.php
@@ -9,6 +9,7 @@ use App\Http\Resources\Org\OrgCollection;
 use App\Http\Resources\Org\OrgResource;
 use App\Models\Org;
 use App\Traits\IncludeRelationships;
+use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 
@@ -16,8 +17,9 @@ class OrgController extends Controller
 {
     use IncludeRelationships;
 
-    public function index(OrgFilter $filter): OrgCollection
+    public function index(OrgFilter $filter, Request $request): OrgCollection
     {
+        $this->authorize('viewAny', Org::class);        
         $org = Org::filter($filter);
 
         return new OrgCollection(
@@ -27,6 +29,7 @@ class OrgController extends Controller
 
     public function store(StoreOrgRequest $request): OrgResource
     {
+        $this->authorize('create', Org::class);
         $validated = $request->validated();
         $org = Org::create($validated);
 
@@ -35,14 +38,18 @@ class OrgController extends Controller
 
     public function show(string $id): OrgResource
     {
-        $org = Org::query();
-        $this->applyIncludes($org, request());
+        
+        $orgQuery = Org::query();
+        $this->applyIncludes($orgQuery, request());
+        $org = $orgQuery->findOrFail($id);
+        $this->authorize('view', $org);
 
-        return new OrgResource($org->findOrFail($id));
+        return new OrgResource($org);
     }
 
     public function update(UpdateOrgRequest $request, Org $org): OrgResource
     {
+        $this->authorize('update', $org);
         $validated = $request->validated();
         $org->update($validated);
 
@@ -51,6 +58,7 @@ class OrgController extends Controller
 
     public function destroy(Org $org): Response
     {
+        $this->authorize('delete', $org);
         $org->deleted_by = Auth::id();
         $org->save();
         $org->delete();
